requirements-dev.txt
# Development dependencies for Greene County Property Finder
# Install with: pip install -r requirements-dev.txt

# Include base requirements
-r requirements.txt

# =============================================================================
# Testing
# =============================================================================
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
coverage>=7.0.0

# =============================================================================
# Code Quality
# =============================================================================
# Formatting
black>=23.0.0
isort>=5.12.0

# Linting
flake8>=6.0.0
flake8-docstrings>=1.7.0
flake8-bugbear>=23.0.0
pylint>=2.17.0

# Type checking
mypy>=1.0.0
pandas-stubs>=2.0.0
types-requests>=2.31.0

# =============================================================================
# Documentation
# =============================================================================
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
myst-parser>=1.0.0

# =============================================================================
# Development Tools
# =============================================================================
# Pre-commit hooks
pre-commit>=3.0.0

# Debugging
ipdb>=0.13.0
ipython>=8.0.0

# Profiling
memory-profiler>=0.61.0
line-profiler>=4.0.0

# =============================================================================
# Build Tools
# =============================================================================
build>=0.10.0
twine>=4.0.0
wheel>=0.40.0

# =============================================================================
# Jupyter (Optional - for exploration)
# =============================================================================
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
